requirements.txt
azure-cli
Django
python-dotenv
whitenoise
